Update expo module scripts
