+++ react-native/client/example/android/app/src/main/res/values/colors.xml
@@ -1,0 +1,4 @@
+<resources>
+  <color name="colorPrimary">#023c69</color>
+  <color name="colorPrimaryDark">#ffffff</color>
+</resources>
