+++ react-native/client/example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">CBW Mobile SDK Example</string>
+  <string name="app_name">example</string>
 </resources>
