+++ react-native/client/example/ios/Podfile.properties.json
@@ -1,0 +1,3 @@
+{
+  "expo.jsEngine": "hermes"
+}
