+++ react-native/client/example/ios/example/AppDelegate.h
@@ -1,9 +1,7 @@
-#import <React/RCTBridgeDelegate.h>
-#import <Expo/Expo.h>
+#import <RCTAppDelegate.h>
 #import <UIKit/UIKit.h>
+#import <Expo/Expo.h>
 
-@interface AppDelegate : EXAppDelegateWrapper <UIApplicationDelegate, RCTBridgeDelegate>
-
-@property (nonatomic, strong) UIWindow *window;
+@interface AppDelegate : EXAppDelegateWrapper
 
 @end
