+++ react-native/client/example/ios/example/Images.xcassets/Contents.json
@@ -1,6 +1,6 @@
 {
   "info" : {
     "version" : 1,
-    "author" : "xcode"
+    "author" : "expo"
   }
 }
