+++ react-native/client/example/ios/example/noop-file.swift
@@ -1,0 +1,4 @@
+//
+// @generated
+// A blank Swift file must be created for native modules with Swift files to work correctly.
+//
