+++ react-native/client/example/package.json
@@ -10,10 +10,13 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "expo": "^46.0.16",
-    "react": "^18.2.0",
-    "react-native": "^0.70.4",
-    "react-native-mmkv": "^2.4.3"
+    "expo": "~48.0.6",
+    "expo-build-properties": "~0.5.1",
+    "expo-splash-screen": "~0.18.1",
+    "expo-status-bar": "~1.4.4",
+    "react": "18.2.0",
+    "react-native": "0.71.3",
+    "react-native-mmkv": "^2.6.3"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
@@ -21,7 +24,7 @@
     "@react-native-community/eslint-config": "^2.0.0",
     "babel-jest": "^26.6.3",
     "eslint": "^7.32.0",
-    "jest": "^26.6.3",
+    "jest": "^29.2.1",
     "metro-react-native-babel-preset": "^0.72.1",
     "react-test-renderer": "18.1.0"
   },
