+++ react-native/client/package.json
@@ -35,8 +35,10 @@
     "react-native-mmkv": "2.4.3"
   },
   "devDependencies": {
-    "expo-module-scripts": "^2.1.1",
-    "expo-modules-core": "^0.11.8"
+    "@types/react": "^18.0.25",
+    "@types/react-native": "^0.70.6",
+    "expo-module-scripts": "^3.0.4",
+    "expo-modules-core": "^1.1.0"
   },
   "peerDependencies": {
     "expo": "*",
